test commit

+++ LifeInYou/DataManager/DatabaseManager.swift
@@ -132,7 +132,6 @@ class DatabaseManager {
     }
 
         
-        
     func editCollectionTask(by user: User,in task: String, oldTask: String, newTask: String, completionDate: String, isComplete: Bool) {
         let oldTask = Task(name: oldTask)
         let newTask = Task(name: newTask)
@@ -140,19 +139,25 @@ class DatabaseManager {
         let task = TaskList(name: task)
 
         DispatchQueue.main.async {
-            self.db.collection("users").document("\(user.uid)").collection("taskList").document("\(task.name)").collection("tasks").document("\(oldTask.name)").delete()
-            self.db.collection("users").document("\(user.uid)").collection("taskList").document("\(task.name)").collection("tasks").document("\(newTask.name)").updateData([
-                "task" : newTask.name,
-                "completionDate" : completionDate.completionDate,
-                "date" : newTask.date,
-                "isComplete" : isComplete,
-                "imageURL" : newTask.imageURL
-            ])
+            self.db.collection("users").document("\(user.uid)").collection("taskList").document("\(task.name)").collection("tasks").document("\(oldTask.name)").getDocument { doc, error in
+                if error == nil {
+                    let data = doc?.data()
+                    self.db.collection("users").document("\(user.uid)").collection("taskList").document("\(task.name)").collection("tasks").document("\(newTask.name)").setData(data ?? ["" : ""])
+                    self.db.collection("users").document("\(user.uid)").collection("taskList").document("\(task.name)").collection("tasks").document("\(newTask.name)").updateData([
+                        "task" : newTask.name,
+                        "completionDate" : completionDate.completionDate,
+                        "isComplete" : isComplete,
+                        "imageURL" : newTask.imageURL
+                    ])
+                    
+                    self.db.collection("users").document("\(user.uid)").collection("taskList").document("\(task.name)").collection("tasks").document("\(oldTask.name)").delete()
+                    
+                }
+            }
         }
     }
     
- 
-        
+
     
 //MARK: - Photo Methods
     func upload(user: String, fromTask: String, task: String, photo: UIImage, completion: @escaping (Result<URL, Error>) -> Void) {
